Release v0.2.0 (#50)
* Added `MockBackend.rows("col1", "col2")[(...), (...)]` helper
([#49](https://github.com/databrickslabs/lsql/issues/49)). In this
release, we have added a new helper method `MockBackend.rows("col1",
"col2")[(...), (...)]` to simplify testing with `MockBackend`. This
method allows for the creation of rows using a more concise syntax,
taking in the column names and a list of values to be used for each
column, and returning a list of `Row` objects with the specified columns
and values. Additionally, a `__eq__` method has been introduced to check
if two rows are equal by converting the rows to dictionaries using the
existing `as_dict` method and comparing them. The `__contains__` method
has also been modified to improve the behavior of the `in` keyword when
used with rows, ensuring columns can be checked for membership in the
row in a more intuitive and predictable manner. These changes make it
easier to test and work with `MockBackend`, improving overall quality
and maintainability of the project.