Release v0.2.3 (#60)
* Added support for common parameters in StatementExecutionBackend
([#59](https://github.com/databrickslabs/lsql/issues/59)). The
`StatementExecutionBackend` class in the `databricks.labs.lsql`
package's `backends.py` file now supports the passing of common
parameters through keyword arguments (kwargs). This enhancement allows
for greater customization and flexibility in the backend's operation, as
the kwargs are passed to the `StatementExecutionExt` constructor. This
change empowers users to control the behavior of the backend, making it
more adaptable to various use cases. The key modification in this commit
is the addition of the `**kwargs` parameter in the constructor signature
and passing it to `StatementExecutionExt`, with no changes made to any
methods within the class.