Release v0.2.5 (#73)
* Fixed PyPI badge
([#72](https://github.com/databrickslabs/lsql/issues/72)). In this
release, we have implemented a fix to the PyPI badge in the README file
of our open-source library. The PyPI badge displays the version of the
package and serves as a quick reference for users. This fix ensures the
accuracy and proper functioning of the badge, without involving any
changes to the functionality or methods within the project. Software
engineers can be assured that this update is limited to the README file,
specifically the PyPI badge, and will not affect the overall
functionality of the library.
* Fixed `no-cheat` check
([#71](https://github.com/databrickslabs/lsql/issues/71)). In this
release, we have made improvements to the `no-cheat` verification
process for new code. Previously, the check for disabling the linter was
prone to false positives when the string '# pylint: disable' appeared
for reasons other than disabling the linter. The updated code now
includes an additional filter to exclude the string `CHEAT` from the
search, and the number of characters in the output is counted using the
`wc -c` command. If the count is not zero, the script will terminate
with an error message. This change enhances the accuracy of the
`no-cheat` check, ensuring that the linter is being used correctly and
that all new code meets our quality standards.
* Removed upper bound on `sqlglot` dependency
([#70](https://github.com/databrickslabs/lsql/issues/70)). In this
update, we have removed the upper bound on the `sqlglot` dependency
version in the project's `pyproject.toml` file. Previously, the version
constraint required `sqlglot` to be at least 22.3.1 but less than
22.5.0. With this modification, there will be no upper limit, enabling
the project to utilize any version greater than or equal to 22.3.1. This
change provides the project with the flexibility to take advantage of
future bug fixes, performance improvements, and new features available
in newer `sqlglot` package versions. Developers should thoroughly test
the updated package version to ensure compatibility with the existing
codebase.