Release v0.3.0 (#76)
* Added support for `save_table(..., mode="overwrite")` to
`StatementExecutionBackend`
([#74](https://github.com/databrickslabs/lsql/issues/74)). In this
release, we've added support for overwriting a table when saving data
using the `save_table` method in the `StatementExecutionBackend`.
Previously, attempting to use the `overwrite` mode would raise a
`NotImplementedError`. Now, when this mode is specified, the method
first truncates the table before inserting the new rows. The truncation
is done using the `execute` method to run a `TRUNCATE TABLE` SQL
command. Additionally, we've added a new integration test,
`test_overwrite`, to the `test_deployment.py` file to verify the new
`overwrite` mode functionality. A new option, `mode="overwrite"`, has
been added to the `save_table` method, allowing for the existing data in
the table to be deleted and replaced with the new data being written.
We've also added two new test cases,
`test_statement_execution_backend_save_table_overwrite_empty_table` and
`test_mock_backend_overwrite`, to verify the new functionality. It's
important to note that the method signature has been updated to include
a default value for the `mode` parameter, setting it to `append` by
default. This change does not affect the functionality and only provides
a more convenient default behavior for users of the method.