Release v0.4.0 (#91)
* Added catalog and schema parameters to execute and fetch
([#90](https://github.com/databrickslabs/lsql/issues/90)). In this
release, we have added optional `catalog` and `schema` parameters to the
`execute` and `fetch` methods in the `SqlBackend` abstract base class,
allowing for more flexibility when executing SQL statements in specific
catalogs and schemas. These updates include new method signatures and
their respective implementations in the `SparkSqlBackend` and
`DatabricksSqlBackend` classes. The new parameters control the catalog
and schema used by the `SparkSession` instance in the `SparkSqlBackend`
class and the `SqlClient` instance in the `DatabricksSqlBackend` class.
This enhancement enables better functionality in multi-catalog and
multi-schema environments. Additionally, this change comes with unit
tests and integration tests to ensure proper functionality. The new
parameters can be used when calling the `execute` and `fetch` methods.
For example, with a `SparkSqlBackend` instance `spark_backend`, you can
execute a SQL statement in a specific catalog and schema with the
following code: `spark_backend.execute("SELECT * FROM my_table",
catalog="my_catalog", schema="my_schema")`. Similarly, the `fetch`
method can also be used with the new parameters.