Release v0.4.2 (#96)
* Added more `NotFound` error type
([#94](https://github.com/databrickslabs/lsql/issues/94)). In the latest
update, the `core.py` file in the `databricks/labs/lsql` package has
undergone enhancements to the error handling functionality. The
`_raise_if_needed` function has been modified to raise a `NotFound`
error when the error message includes the phrase "does not exist". This
update enables the system to categorize specific SQL query errors as
`NotFound` error messages, thereby improving the overall error handling
and reporting capabilities. This change was a collaborative effort, as
indicated by the co-authored-by statement in the commit.