Release v0.4.3 (#101)
* Bump actions/checkout from 4.1.2 to 4.1.3
([#97](https://github.com/databrickslabs/lsql/issues/97)). The
`actions/checkout` dependency has been updated from version 4.1.2 to
4.1.3 in the `update-main-version.yml` file. This new version includes a
check to verify the git version before attempting to disable
`sparse-checkout`, and adds an SSH user parameter to improve
functionality and compatibility. The release notes and CHANGELOG.md file
provide detailed information on the specific changes and improvements.
The pull request also includes a detailed commit history and links to
corresponding issues and pull requests on GitHub for transparency. You
can review and merge the pull request to update the `actions/checkout`
dependency in your project.
* Maintain PySpark compatibility for databricks.labs.lsql.core.Row
([#99](https://github.com/databrickslabs/lsql/issues/99)). In this
release, we have added a new method `asDict` to the `Row` class in the
`databricks.labs.lsql.core` module to maintain compatibility with
PySpark. This method returns a dictionary representation of the `Row`
object, with keys corresponding to column names and values corresponding
to the values in each column. Additionally, we have modified the `fetch`
function in the `backends.py` file to return `Row` objects of
`pyspark.sql` when using `self._spark.sql(sql).collect()`. This change
is temporary and marked with a `TODO` comment, indicating that it will
be addressed in the future. We have also added error handling code in
the `fetch` function to ensure the function operates as expected. The
`asDict` method in this implementation simply calls the existing
`as_dict` method, meaning the behavior of the `asDict` method is
identical to the `as_dict` method. The `as_dict` method returns a
dictionary representation of the `Row` object, with keys corresponding
to column names and values corresponding to the values in each column.
The optional `recursive` argument in the `asDict` method, when set to
`True`, enables recursive conversion of nested `Row` objects to nested
dictionaries. However, this behavior is not currently implemented, and
the `recursive` argument is always `False` by default.

Dependency updates:

* Bump actions/checkout from 4.1.2 to 4.1.3
([#97](https://github.com/databrickslabs/lsql/pull/97)).