Release v0.8.0 (#250)
* Removed deploy_dashboard method
([#240](https://github.com/databrickslabs/lsql/issues/240)). In this
release, the `deploy_dashboard` method has been removed from the
`dashboards.py` file and the legacy deployment method has been
deprecated. The `deploy_dashboard` method was previously used to deploy
a dashboard to a workspace, but it has been replaced with the `create`
method of the `lakeview` attribute of the WorkspaceClient object.
Additionally, the `test_dashboards_creates_dashboard_via_legacy_method`
method has been removed. A new test has been added to ensure that the
`deploy_dashboard` method is no longer being used, utilizing the
`deprecated_call` function from pytest to verify that calling the method
raises a deprecation warning. This change simplifies the code and
improves the overall design of the system, resolving issue
[#232](https://github.com/databrickslabs/lsql/issues/232). The
`_with_better_names` method and `create_dashboard` method remain
unchanged.
* Skip test that fails due to insufficient permission to create schema
([#248](https://github.com/databrickslabs/lsql/issues/248)). A new test
function, `test_dashboards_creates_dashboard_with_replace_database`, has
been added to the open-source library, but it is currently marked to be
skipped due to missing permissions to create a schema. This function
creates an instance of the `Dashboards` class with the `ws` parameter,
creates a dashboard using the `make_dashboard` function, and performs
various actions using the created dashboard, as well as functions such
as `tmp_path` and `sql_backend`. This test function aims to ensure that
the `Dashboards` class functions as expected when creating a dashboard
with a replaced database. Once the necessary permissions for creating a
schema are acquired, this test function can be enabled for further
testing and validation.
* Updates to use the Databricks Python sdk 0.30.0
([#247](https://github.com/databrickslabs/lsql/issues/247)). In this
release, we have updated the project to use Databricks Python SDK
version 0.30.0. This update includes changes to the `execute` and
`fetch_value` functions, which now use the new `StatementResponse` type
instead of `ExecuteStatementResponse`. A conditional import statement
has been added to maintain compatibility with both Databricks SDK
versions 0.30.0 and below. The `execute` function now raises
`TimeoutError` when the specified timeout is greater than 50 seconds and
the statement execution hasn't finished. Additionally, the `fetch_value`
function has been updated to handle the case when the `execute` function
returns `None`. The unit test file `test_backends.py` has also been
updated to reflect these changes, with multiple test functions now using
the `StatementResponse` class instead of `ExecuteStatementResponse`.
These changes improve the system's compatibility with the latest version
of the Databricks SDK, ensuring that the core functionality of the SDK
continues to work as expected.