Release v0.9.3 (#267)
* Added documentation for exclude flag
([#265](https://github.com/databrickslabs/lsql/issues/265)). A new
`exclude` flag has been added to the configuration file for our lab
tool, allowing users to specify a path to exclude from formatting during
lab execution. This release also includes corrections to grammatical
errors in the descriptions of existing flags related to catalog and
database settings, such as updating `seperated` to "separate".
Additionally, the flag descriptions for `publish` and `open-browser`
have been updated for clarification: `publish` now clearly controls
whether the dashboard is published after creation, while `open-browser`
controls whether the dashboard is opened in a web browser. These changes
are aimed at improving user experience and ease of use for our lab tool.
* Fixed dataclass field type in _row_to_sql
([#266](https://github.com/databrickslabs/lsql/issues/266)). In this
release, we have addressed an issue related to
[#257](https://github.com/databrickslabs/lsql/issues/257) by fixing the
dataclass field type in the `_row_to_sql` method of the `backends.py`
file. Additionally, we have made updates to the `_schema_for` method to
use a new `_field_type` class method. This change resolves a rare
problem where the `field.type` is a string instead of a type and ensures
compatibility with a pull request from an external repository
(<https://github.com/databrickslabs/ucx/pull/2526>). The new
`_field_type` method attempts to load the type from `__builtins__` if
it's a string and logs a warning if it fails. The `_row_to_sql` method
now consistently uses the `_field_type` method to get the field type.
This ensures that the library functions seamlessly and consistently,
avoiding any potential issues in the future.