New example with bar chart in dashboards-as-code (#281)
Resolves #116 

To add support for other Widgets like Bar, Pivot, Area, etc., we need to
add extra information to TileMetadata to render them properly. These
widgets have different attributes, and adding new metadata properties
for each of them may bloat the TileMetadata structure.

Should we just use the existing metadata overrides feature to support
these new Widget types or extend the current implementation? To test if
overrides are sufficient to create new Widgets, an integration test has
been added to create a bar chart. Attached is the screenshot for the
dashboard with a bar chart. The overrides approach works, as shown in
the screenshot.

@nfx @JCZuurmond Please provide your feedback on how you want to go
forward with this.


![image](https://github.com/user-attachments/assets/a6f03b06-ac64-4993-b802-eb3538067047)