Release v0.12.0 (#293)
* Added method to detect rows are written to the `MockBackend`
([#292](https://github.com/databrickslabs/lsql/issues/292)). In this
commit, the `MockBackend` class in the 'backends.py' file has been
updated with a new method, 'has_rows_written_for', which allows for
differentiation between a table that has never been written to and one
with zero rows. This method checks if a specific table has been written
to by iterating over the table stubs in the `_save_table` attribute and
returning `True` if the given full name matches any of the stub full
names. Additionally, the class has been supplemented with the
`rows_written_for` method, which takes a table name and mode as input
and returns a list of rows written to that table in the given mode.
Furthermore, several new test cases have been added to test the
functionality of the `MockBackend` class, including checking if the
`has_rows_written_for` method correctly identifies when there are no
rows written, when there are zero rows written, and when rows are
written after the first and second write operations. These changes
improve the overall testing coverage of the project and aid in testing
the functionality of the `MockBackend` class. The new methods are
accompanied by documentation strings that explain their purpose and
functionality.