Minor improvements for `.save_table(mode="overwrite")` (#298)
This PR includes some updates for `.save_table()` in overwrite mode:

- If no rows are supplied, instead of a no-op the table is truncated.
The mock backend already assumed this but diverged from the concrete
implementations which treated this as a no-op (as when appending). Unit
tests now cover this situation for all backends, and the existing
integration test for the SQL-statement backend has been updated to cover
this.
- The SQL-based backends have a slight optimisation: instead of first
truncating before inserting the truncate is now performed as part of the
insert for the first batch.
- Type hints on the abstract method now match the concrete
implementations.