Deprecate `Row.as_dict()` (#309)
This PR deprecates the `Row.as_dict()` method: this class is intended to
be interchangeable with Spark's `Row`, which doesn't support
`.as_dict()` method. Having it present leads to subtle bugs where code
works when the execution-based backends are being used but fails when
the Spark-based backend is used instead. Clients should instead use
`.asDict()`.

To signal the deprecation we use Python's warnings mechanism (including
the use of the new annotation in Python 3.13, to help with static code
analysis).