Release v0.14.1 (#352)
* Changes to work with Databricks SDK `v0.38.0`
([#350](https://github.com/databrickslabs/lsql/issues/350)). In this
release, we have upgraded the Databricks SDK to version 0.38.0 from
version 0.37.0 to ensure compatibility with the latest SDK and address
several issues. The update includes changes to make the code compatible
with the new SDK version, removing the need for `.as_dict()` method
calls when creating or updating dashboards and utilizing a
`sdk_dashboard` variable for interacting with the Databricks workspace.
We also updated the dependencies to "databricks-labs-blueprint[yaml]"
package version greater than or equal to 0.4.2 and `sqlglot` package
version greater than or equal to 22.3.1. The `test_core.py` file has
been updated to address multiple issues
([#349](https://github.com/databrickslabs/lsql/issues/349) to
[#332](https://github.com/databrickslabs/lsql/issues/332)) related to
the Databricks SDK and the `test_dashboards.py` file has been revised to
work with the new SDK version. These changes improve integration with
Databricks' lakeview dashboards, simplify the code, and ensure
compatibility with the latest SDK version, resolving issues
[#349](https://github.com/databrickslabs/lsql/issues/349) to
[#332](https://github.com/databrickslabs/lsql/issues/332).
* Specify the minimum required version of `databricks-sdk` as 0.37.0
([#331](https://github.com/databrickslabs/lsql/issues/331)). In this
release, we have updated the minimum required version of the
`databricks-sdk` package to 0.37.0 from 0.29.0 in the `pyproject.toml`
file to ensure compatibility with the latest version. This change was
made necessary due to updates made in issue
[#320](https://github.com/databrickslabs/lsql/issues/320). To
accommodate any patch release of `databricks-sdk` with a major and minor
version of 0.37, we have updated the dependency constraint to use the
`~=` operator, resolving issue
[#330](https://github.com/databrickslabs/lsql/issues/330). These changes
are intended to enhance the compatibility and stability of our software.