Release v0.15.1 (#368)
* remove duplicate changelog entry
([#366](https://github.com/databrickslabs/lsql/issues/366)). In this
update, we've improved resource name validation in the `ucx` project,
first introduced in version 0.14.2. The validation now restricts
resource names to alphanumeric characters, hyphens, and underscores,
addressing usability issues and preventing problems caused by special
characters. A new internal method, `_is_valid_resource_name`, checks if
a name is valid according to the defined pattern. The `TileMetadata`
class has been updated to ensure its `id` attribute follows the new
validation rules, and `Tile`, `Section`, and `Dashboard` classes now
call the `validate` method of the `TileMetadata` instance if it exists.
These changes promote consistency and correctness in dashboard
resources, simplifying management and interaction for users. As a part
of this commit, we also removed a duplicate changelog entry related to
this feature that was previously present in version 0.15.0.