Limit `sqlglot` to releases earlier than 26.30.0. (#410)
The 26.30.0 release of sqlglot introduced a breaking change that affects
our tests; this PR is a quick-fix to prevent that version from being
used.

This PR also includes type-hinting fixes that newer versions of mypy
need, along with accompanying fixes for issues that the improved
type-hints expose.

For now this is intended to:

 - Unblock databrickslabs/blueprint#248.
 - Supersede #409.